rename file from index.ts -> vcs.ts
