+++ packages/insomnia-app/app/sync/vcs/__tests__/vcs.test.ts
@@ -1,4 +1,4 @@
-import VCS from '../index';
+import VCS from '../vcs';
 import MemoryDriver from '../../store/drivers/memory-driver';
 import { describeChanges } from '../util';
 import { globalBeforeEach } from '../../../__jest__/before-each';
