+++ packages/insomnia-app/app/ui/components/wrapper.tsx
@@ -70,7 +70,7 @@ import type { ClientCertificate } from '../../models/client-certificate';
 import MoveRequestGroupModal from './modals/move-request-group-modal';
 import AddKeyCombinationModal from './modals/add-key-combination-modal';
 import ExportRequestsModal from './modals/export-requests-modal';
-import VCS from '../../sync/vcs';
+import VCS from '../../sync/vcs/vcs';
 import type { StatusCandidate } from '../../sync/types';
 import type { RequestMeta } from '../../models/request-meta';
 import type { RequestVersion } from '../../models/request-version';
